Create minimum prefab for state machine tests
