+++ Tests/Editor/DummyStateMachine.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 53294ae9beff2d14db14a191d7b19161
+guid: af9d020e6d232fc499c6cc7e3a34de8f
 PrefabImporter:
   externalObjects: {}
   userData:
